requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.1.1
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-cloud-bigquery==3.11.0
matplotlib==3.8.0
torch==2.0.1
